hydra: gate mixed-XOR GE on a WORK estimate, not nvars (Ascon stall)
A live 2026 run sat 10+ minutes silently in the xor stage on
asconhashv12_opt64 (158719 vars, 373656 clauses).  Cause: recover_xors
finds ~79k arity-3 XORs (Ascon's linear diffusion layer) but the formula
is MIXED (the nonlinear chi clauses remain), and the instance is UNDER
the old nvars>200k cap — so dense GF(2) elimination ran.  Unlike
structured Tseitin (banded, low fill-in: n400 at 319k vars = 580s), a
crypto diffusion matrix fills in catastrophically, so cost tracks
XOR-count × bitset-width, not nvars.  And it's wasted: the linear layer
alone can't decide a hash preimage → GE returns Indeterminate after the
grind.

solve_xor_system's cap is now a WORK estimate `n_xors × (nvars/64)`
(u64, saturating) instead of nvars; hydra + eff pass 50_000_000.  That
skips both 2026 Ascon instances (work ~1.2-2.0e8) and the 487k-var queens
case (~3.4e9), while every small mixed simplification still runs.
PURE-XOR formulas remain ungated (n400 still decides + certifies).  The
skip now prints in <1ms ("N XOR(s) recovered, no usable consequence
(0.6ms)") instead of stalling silently.  +1 unit test
(mixed_work_cap_skips_but_pure_xor_is_never_gated).

NOT YET BUILT INTO target/release/sat — prepared for a rebuild after the
live 2026 run finishes:  cargo build --release --bin sat

+++ examples/xordbg.rs
@@ -35,7 +35,7 @@ fn main() {
     for x in xors.iter().take(8) {
         println!("  sample: vars={:?} rhs={}", &x.vars[..x.vars.len().min(6)], x.rhs);
     }
-    match solve_xor_system(nvars, &clauses, usize::MAX) {
+    match solve_xor_system(nvars, &clauses, u64::MAX) {
         XorGaussResult::Unsat { by_bcp } => println!("solve_xor_system: UNSAT (by_bcp={})", by_bcp),
         XorGaussResult::Sat(_) => println!("solve_xor_system: SAT"),
         XorGaussResult::Simplified { recovered, forced_count, .. } =>
